Initial commit
Options reads information from Yahoo! Finance for a given ticker(s) and
extracts from it the values of interest. Namely at this point, those
would be the nearest out of the money put and what it is selling for as
a ratio to the underlying stock price.

Can output in a formatted table or to CSV.