Refactor
Instead of using the JSON structure we get from symbol() and having to
constantly error check it, first scrub it and convert it over to an
internal data structure (security).

This saves a lot of error checking and it makes the flow of the program
simpler to follow. If the data comes in dirty we can abort quickly.