Use a common options file
It turns out that the scan for which stocks have options is low yield.
Running the scan gets about a 30% false negative rate. So, instead of
having the scan run fully each time we get a new USE file, maintain a
known list of optionable stocks and only check on stocks not in that
list.