Be pickier about which options to consider
If an option does not have weekly (or more frequent) expirations then do
not consider it. If we sell puts and get assigned we will want to be
able to 'farm' the shares every week for income until we can unload
them. We don't want to have to sit on shares for multple weeks, not
being able to monetize them.