Add CPU profiling, make speed improvements
Add CPU profiling.

It turns out we were writing back to the cache each time we read from
the cache. Stop doing that.

The JSON unmarshal code was slow. Replace it with faster code, which
also happens to be simpler.

When displaying the list of found options, change the semantices of
`-expiration` to mean expiration dates between now and then, rather than
just that date alone. This allows us to see more choices, like SPY which
expire quite frequently.