Calculate the 'Call Spread'
If we get assigned on a put then we will need to sell calls until the
shares are assigned away. This means that we want to favor securities
that we can make some money selling calls on. Rate each security on
how far out there are strikes that still have non-zero bids.