Sleep in small increments
time.Sleep() gets confused if the computer hibernates while it is
sleeping. It seems that the time the computer spends in hibernation is
not credited towards the elapsed total time. So, sleeping for 12 hours
would normally take 12 hours but if the computer hibernates for 10 hours
then it would actually take 22 total hours.

Sleep in small increments to minimize this effect.